+++ src/Vector/BLF/LinMessage2.cpp
@@ -29,6 +29,7 @@ LinMessage2::LinMessage2() :
 }
 
 void LinMessage2::read(AbstractFile & is) {
+    apiMajor = 1;
     ObjectHeader::read(is);
     LinDatabyteTimestampEvent::read(is);
     is.read(reinterpret_cast<char *>(data.data()), static_cast<std::streamsize>(data.size()));
@@ -44,18 +45,18 @@ void LinMessage2::read(AbstractFile & is) {
     is.read(reinterpret_cast<char *>(&reservedLinMessage2), sizeof(reservedLinMessage2));
 
     /* the following variables are only available in Version 2 and above */
-    /*if (objectVersion < 0)*/ // Vector bug: Shouldn't this be < 1?
-    /*    return;*/
-
-    is.read(reinterpret_cast<char *>(&respBaudrate), sizeof(respBaudrate));
+    if (objectSize > calculateObjectSize()) {
+        apiMajor = 2;
+        is.read(reinterpret_cast<char *>(&respBaudrate), sizeof(respBaudrate));
+    }
 
     /* the following variables are only available in Version 3 and above */
-    if (objectVersion < 1)   // Vector bug: Shouldn't this be < 2?
-        return;
-
-    is.read(reinterpret_cast<char *>(&exactHeaderBaudrate), sizeof(exactHeaderBaudrate));
-    is.read(reinterpret_cast<char *>(&earlyStopbitOffset), sizeof(earlyStopbitOffset));
-    is.read(reinterpret_cast<char *>(&earlyStopbitOffsetResponse), sizeof(earlyStopbitOffsetResponse));
+    if (objectSize > calculateObjectSize()) {
+        apiMajor = 3;
+        is.read(reinterpret_cast<char *>(&exactHeaderBaudrate), sizeof(exactHeaderBaudrate));
+        is.read(reinterpret_cast<char *>(&earlyStopbitOffset), sizeof(earlyStopbitOffset));
+        is.read(reinterpret_cast<char *>(&earlyStopbitOffsetResponse), sizeof(earlyStopbitOffsetResponse));
+    }
 
     // @note might be extended in future versions
 }
@@ -76,15 +77,13 @@ void LinMessage2::write(AbstractFile & os) {
     os.write(reinterpret_cast<char *>(&reservedLinMessage2), sizeof(reservedLinMessage2));
 
     /* the following variables are only available in Version 2 and above */
-    /*if (objectVersion < 0)*/ // Vector bug: Shouldn't this be < 1?
-    /*    return;*/
-
+    if (apiMajor < 2)
+        return;
     os.write(reinterpret_cast<char *>(&respBaudrate), sizeof(respBaudrate));
 
     /* the following variables are only available in Version 3 and above */
-    if (objectVersion < 1)   // Vector bug: Shouldn't this be < 2?
+    if (apiMajor < 3)
         return;
-
     os.write(reinterpret_cast<char *>(&exactHeaderBaudrate), sizeof(exactHeaderBaudrate));
     os.write(reinterpret_cast<char *>(&earlyStopbitOffset), sizeof(earlyStopbitOffset));
     os.write(reinterpret_cast<char *>(&earlyStopbitOffsetResponse), sizeof(earlyStopbitOffsetResponse));
@@ -106,13 +105,19 @@ uint32_t LinMessage2::calculateObjectSize() const {
         sizeof(reservedLinMessage1) +
         sizeof(reservedLinMessage2);
 
-    /*if (objectVersion >= 0)*/ // Vector bug: Shouldn't this be >= 1?
-    size += sizeof(respBaudrate);
+    /* the following variables are only available in Version 2 and above */
+    if (apiMajor < 2)
+        return size;
+    size +=
+        sizeof(respBaudrate);
 
-    if (objectVersion >= 1) // Vector bug: Shouldn't this be >= 2?
-        size += sizeof(exactHeaderBaudrate) +
-                sizeof(earlyStopbitOffset) +
-                sizeof(earlyStopbitOffsetResponse);
+    /* the following variables are only available in Version 3 and above */
+    if (apiMajor < 3)
+        return size;
+    size +=
+        sizeof(exactHeaderBaudrate) +
+        sizeof(earlyStopbitOffset) +
+        sizeof(earlyStopbitOffsetResponse);
 
     return size;
 }