+++ CHANGELOG.md
@@ -5,8 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - New Types: CanSettingChanged, DistributedObjectMember, AttributeEvent
-### Fixed
+### Changed
 - Drop VectorTypes.h and use std types instead.
+### Fixed
+- Fix issues with newer BLF in CanFdMessage, CanMessage2, EthernetFrame, LinMessage2
 
 ## [2.3.1] - 2020-07-16
 ### Fixed